Added a `Picture` component for use with `Node.picture`.
It creates a set of `.sources` using all the supplied `Attribute`s relying the behavior of `Attribute` to not render null strings and/or `.empty` nodes.

I tried to use the `.if` syntax, but got a compiler error. I am leaving the commented code in case there is a better way to implement this.